apps/backend/requirements/base.txt
Django==5.0.6
psycopg[binary]==3.1.18
dj-database-url==2.2.0
celery==5.4.0
redis==5.0.8
djangorestframework==3.15.2
